% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixEncoding}
\alias{fixEncoding}
\title{Sets encoding to utf8 for data}
\usage{
fixEncoding(data)
}
\description{
internal function used by rccShiny to "clean" data to utf8 encoding.
}
\author{
Fredrik Sandin, RCC Mellansverige
}
\keyword{internal}
