% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fLinePlot.R
\name{fLinePlot}
\alias{fLinePlot}
\title{Creates trend plot}
\usage{
fLinePlot(
  x = NULL,
  y = NULL,
  legend = NULL,
  legend_pos = "bottom",
  legend_pch = 15,
  legend_ncol = NULL,
  legend_textwidth = NULL,
  col = NULL,
  stacked_area = FALSE,
  linewidth = 4,
  linetype = "l",
  markers = TRUE,
  x_lim = NULL,
  y_lim = NULL,
  x_by = NULL,
  y_by = NULL,
  x_ticks_labels = NULL,
  title = NULL,
  subtitle1 = NULL,
  subtitle2 = NULL,
  x_lab = "x",
  y_lab = "y",
  text_cex = 1,
  target_values = NULL,
  target_values_high = NULL,
  target_values_labels = c("Mellannivå av måluppfyllelse",
    "Hög nivå av måluppfyllelse")
)
}
\description{
internal function used by server.R for plot in tab Trend.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\author{
Fredrik Sandin, RCC Mellansverige
}
\keyword{internal}
