% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{npcrPreparePeriodRegionCountyHospitalVariables}
\alias{npcrPreparePeriodRegionCountyHospitalVariables}
\title{Used by NPCR}
\usage{
npcrPreparePeriodRegionCountyHospitalVariables(prefix = "d0", data = NULL,
  language = c("sv", "en"), returnPrivateOthersNames = FALSE)
}
\description{
internal function.
}
\author{
Fredrik Sandin, RCC Uppsala-Örebro
}
