% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc2ColShade.R
\name{rcc2ColShade}
\alias{rcc2ColShade}
\title{rcc2ColShade}
\usage{
rcc2ColShade(col = "#000000", factor = seq(0, 2, 0.1))
}
\description{
internal function.
}
\author{
Fredrik Sandin, RCC Uppsala-Örebro
}
