% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc2PlotInd.R
\name{rcc2PlotInd}
\alias{rcc2PlotInd}
\title{rcc2PlotInd}
\usage{
rcc2PlotInd(
  group = NULL,
  groupHideLessThan = FALSE,
  groupHideLessThanLabel = "(otillräcklig data)",
  groupHideLessThanCell = FALSE,
  groupMaxChars = NULL,
  ind = NULL,
  period = NULL,
  periodMaxN = 99,
  periodIncludeGroupsNotInLatestPeriod = TRUE,
  indType = class(ind),
  indNumeric = indType \%in\% c("difftime", "numeric", "integer"),
  indNumericExcludeNeg = TRUE,
  indNumericPercentiles = c(0.25, 0.5, 0.75),
  indFactorHide = NULL,
  indFactorSortByCols = NULL,
  indFactorShowN = ifelse(!is.null(indFactorHide), TRUE, FALSE),
  indShowPct = ifelse(indType == "factor", FALSE, TRUE),
  indTitle = ifelse(indNumeric, "Median", "Procent"),
  indNCasesTxt = "Antal fall",
  indNCasesOfTxt = "av",
  legendNCol = NULL,
  legendFixedTextWidth = TRUE,
  col = NULL,
  border = TRUE,
  xMax = if (indNumeric) {     NULL } else {     100 },
  xBy = NULL,
  xLab = ifelse(indNumeric, "Median samt kvartilavstånd", "Procent"),
  allLab = "RIKET",
  emphLab = NULL,
  title = NULL,
  subtitle1 = NULL,
  subtitle2 = NULL,
  cexText = 1,
  cexPoint = 2.25,
  targetValues = NULL,
  targetValuesHigh = NULL,
  targetValuesLabels = c("Mellannivå av måluppfyllelse",
    "Hög nivå av måluppfyllelse"),
  funnelplot = FALSE,
  funnelplotProbs = c(0.05, 0.01),
  sort = TRUE,
  subset = NULL,
  subsetLab = "SUBSET",
  extra = NULL,
  extraLab = "EXTRA",
  outputHighchart = FALSE
)
}
\description{
internal function.
}
\author{
Fredrik Sandin, RCC Uppsala-Örebro
}
\keyword{internal}
