% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc2PlotMap.R
\name{rcc2PlotMap}
\alias{rcc2PlotMap}
\title{rcc2PlotMap}
\usage{
rcc2PlotMap(
  value = NULL,
  valueLim = NULL,
  valueOrder = c("Östergötland", "Blekinge", "Dalarna", "Gävleborg", "Gotland",
    "Halland", "Jämtland", "Jönköping", "Kalmar", "Kronoberg", "Norrbotten", "Orebro",
    "Södermanland", "Skåne", "Stockholm", "Uppsala", "Värmland", "Västerbotten",
    "Västernorrland", "Västmanland", "Västra Götaland"),
  valueOrderReturn = FALSE,
  legend = "",
  col = NULL,
  colBorder = "#FFFFFF",
  cexText = 1,
  title = NULL,
  subtitle1 = NULL,
  subtitle2 = NULL,
  nDec = 0,
  rdsPath = "./data/",
  outputHighchart = FALSE
)
}
\description{
internal function.
}
\author{
Fredrik Sandin, RCC Mellansverige
}
\keyword{internal}
