% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccShiny-package.R
\docType{package}
\name{rccShiny-package}
\alias{rccShiny-package}
\alias{_PACKAGE}
\title{rccShiny: Shiny apps for RCC}
\description{
Creates shiny apps used as a complement to the annual reports from the cancer quality registries in Sweden.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cancercentrum.bitbucket.io/rccshiny/}
  \item \url{http://bitbucket.org/cancercentrum/rccshiny}
}

}
\author{
\strong{Maintainer}: Fredrik Sandin \email{fredrik.sandin@rccuppsalaorebro.se}

Other contributors:
\itemize{
  \item Lina Benson [contributor]
  \item Erik Lindberg [contributor]
  \item John Lövrot [contributor]
  \item Chenyang Zhang [contributor]
  \item RCC [copyright holder, funder]
}

}
\keyword{internal}
