% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccShiny2.R
\name{rccShiny2}
\alias{rccShiny2}
\title{Shiny apps for RCC}
\usage{
rccShiny2(inca = FALSE, incaScript = NULL, incaIncludeList = TRUE,
  folder = "ind", path = getwd(), language = "sv", data = NULL,
  id = NULL, idOverviewLink = NULL, idAuthorisedToView = NULL,
  outcome = "outcome", outcomeNumericExcludeNeg = TRUE,
  outcomeTitle = outcome, textBeforeSubtitle = NULL,
  textAfterSubtitle = NULL, comment = "", description = rep(NA, 3),
  geoUnitsHospital = "sjukhus", geoUnitsHospitalCode = "sjukhuskod",
  geoUnitsHospitalSelected = NULL, geoUnitsCounty = "landsting",
  geoUnitsRegion = "region", geoUnitsPatient = FALSE,
  regionSelection = TRUE, regionLabel = rccShinyTXT(language =
  language)$limitRegion, period = "period", periodDateLevel = "year",
  periodLabel = rccShinyTXT(language = language)$dxYear,
  varOther = NULL, targetValues = NULL, funnelplot = FALSE,
  sortDescending = NULL, propWithinShow = TRUE,
  propWithinUnit = rccShinyTXT(language = language)$propWithinUnit,
  propWithinValue = 30, hideLessThan = 5, gaPath = NULL,
  npcrGroupPrivateOthers = FALSE)
}
\arguments{
\item{inca}{should output be as if on the INCA platform? Default is FALSE.}

\item{incaScript}{script to be run after loading data on the INCA platform. Default is NULL.}

\item{incaIncludeList}{Should the tab with list of patients be included if on the INCA platform? Default is TRUE.}

\item{folder}{name of folder where the results are placed. Default is "ind".}

\item{path}{search path to folder returned by the function. Default is working directory.}

\item{language}{vector giving the language for the app. Possible values are "sv" and "en". Default is "sv". See details.}

\item{data}{data frame containing the variables used when not on the INCA platform.}

\item{id}{optional name of variable in data containing the id of each individual. This is displayed in the list tab if on the INCA platform. Default is NULL.}

\item{idOverviewLink}{optional name of variable in data containing the HTML link to the patient overview on INCA for each individual. This is displayed in the list if on the INCA platform. Default is NULL.}

\item{idAuthorisedToView}{optional name of binary variable in data containing information on whether or not (1 = yes or 0 = no) the user running the app is authorised to see the id and idOverviewLink in the patient list on INCA. If id or idOverviewLink is provided, idAuthorisedToView must also be provided. Default is NULL.}

\item{outcome}{vector with names(s) of variable(s) in data containing the variable(s) to be presented in the app, for example a quality indicator. Variable(s) must be of type logical, factor or numeric. Default is "outcome". Observe that observations with missing values for outcome are not included in the output.}

\item{outcomeNumericExcludeNeg}{should negative values be excluded when presenting a numeric outcome? Particularly relevant for waiting times. Default is TRUE.}

\item{outcomeTitle}{label(s) of the outcome(s) shown in the app. Must be the same length as argument outcome. Default is argument outcome.}

\item{textBeforeSubtitle}{optional text placed before the subtitles in the tabs.}

\item{textAfterSubtitle}{optional text placed after the subtitles in the tabs.}

\item{comment}{optional comment printed under the sidebar panel.}

\item{description}{vector of 3 character strings, or a list of vectors, one for each language, shown in the three subsections in the tab Beskrivning/Description. Default is c(NA, NA, NA).}

\item{geoUnitsHospital}{optional name of variable in data containing hospital names. Variable must be of type character. If NULL or if variable is not found in 'data', hospital is not available as a level of presentation. Default is "sjukhus". At least one geoUnit must be given.}

\item{geoUnitsHospitalCode}{optional name of variable in data containing hospital codes. Variable must be of type numeric. If NULL or if variable is not found in 'data', the list tab can not be displayed. The hospital codes are used to determine which patients to show in the list tab by matching it to the enviromental variable on INCA containing the hospital code of the logged in user. Default is "sjukhuskod".}

\item{geoUnitsHospitalSelected}{optional name of the choice that should initially be selected in the list of hospitals. Variable must be of type character. Default is NULL.}

\item{geoUnitsCounty}{optional name of variable in data containing county codes. Variable must be of type numeric. Can be either county of residence for the patient or the county the hospital belongs to. See details for valid values. If NULL or if variable is not found in 'data', county is not available as a level of presentation. Default is "landsting". At least one geoUnit must be given. To be implemented: Codes for county of hospital are fetched automatically from hospital codes.}

\item{geoUnitsRegion}{optional name of variable in data containing region codes (1=Stockholm, 2=Uppsala-Örebro, 3=Sydöstra, 4=Södra, 5=Västra, 6=Norra, NA=Uppgift saknas). Variable must be of type numeric. Can be either region of residence for the patient or the region the hospital belongs to. If NULL or if variable is not found in 'data', region is not available as a level of presentation. Default is "region". At least one geoUnit must be given. To be implemented: Codes for region of hospital are fetched automatically from hospital codes.}

\item{geoUnitsPatient}{if geoUnitsCounty/geoUnitsRegion is county/region of residence for the patient (LKF). If FALSE and a hospital is chosen by the user in the sidebar panel the output is highlighted for the respective county/region that the hospital belongs to. Default is FALSE.}

\item{regionSelection}{adds a widget to the sidebar panel with the option to show only one region at a time. Default is TRUE.}

\item{regionLabel}{if regionSelection = TRUE label of widget shown in the sidebar panel. Default is "Begränsa till region" or "Limit to region", depending on language.}

\item{period}{name of variable in data containing time periods, for example date or year of diagnosis. Variable must be of type numeric or Date. Default is "period". If period = NULL then no period variable is required and period will not be included anywhere in the Shiny app.}

\item{periodDateLevel}{If the variable in data containing time period is of type Date, how are the time periods going to be grouped? Allowed values are "year and "quarter", with default "year".}

\item{periodLabel}{label for the period widget in the sidebar panal. Default is "Diagnosår", "Year of diagnosis" depending on language.}

\item{varOther}{optional list of variable(s), other than period and geoUnits, to be shown in the sidebar panel. Arguments to the list are: var (name of variable in data), label (label shown over widget in sidebar panel), choices (which values of var should be shown, min, max for continuous variables), selected (which values should be selected when app is launched, default is all avalible values), multiple (should multiple choises be availible, default is TRUE), showInTitle (should selection be displayed in subtitle, default is TRUE). Observe that observations with missing values for varOthers are not included in the output.}

\item{targetValues}{optional vector or list of vectors (one for each outcome) with 1-2 target levels to be plotted in the tabs Jämförelse/Comparison and Trend for outcomes of type logical or numeric. If the outcome is numeric the target levels are shown when "Andel inom..."/"Proportion within..." is selected, and then only for the default propWithinValue.}

\item{funnelplot}{adds a widget to the sidebar panel with the option to show a funnel plot in the tab Jämförelse/Comparison. Only applicaple for dichotomous variables. Default is FALSE.}

\item{sortDescending}{should the bars in tab Jämförelse/Comparison be plotted in descending order? The argument could have the same length as argument outcome, giving different values for each outcome. Default is NULL, which sorts logical outcomes in descending order and continuous outcomes in ascending order.}

\item{propWithinShow}{display the choice "Andel inom..."/"Proportion within..." for numeric outcome(s). Default is TRUE.}

\item{propWithinUnit}{unit shown for numeric outcome when "Andel inom..."/"Proportion within..." is selected. Default is "dagar", "days" depending on language.}

\item{propWithinValue}{vector with default value(s) shown for numeric outcome(s) when "Andel inom..."/"Proportion within..." is selected. The length of the vector should be either 1 or the length of outcome. Default is 30.}

\item{hideLessThan}{value under which groups (cells) are supressed. Default is 5 and all values < 5 are set to 5.}

\item{gaPath}{optional path to Google Analytics .js-file. Default is NULL.}

\item{npcrGroupPrivateOthers}{should private hospitals be grouped when displaying data for the entire country. Applicable for NPCR. Default is FALSE.}
}
\value{
A folder path/sv|en/folder containing: global.R, server.R, ui.R, data/data.RData, docs/description.html.
}
\description{
returns shiny apps that can be used as a complement to the annual reports from the cancer quality registries in Sweden.
}
\details{
Valid values for geoUnitsCounty are:
  \tabular{lll}{
\strong{geoUnitsPatient}   \tab \strong{!geoUnitsPatient} \tab \strong{Text shown}\cr
1\tab 10,11 \tab  Stockholm\cr
3\tab 12 \tab  Uppsala\cr
4\tab 13 \tab  Södermanland\cr
5\tab 21 \tab  Östergötland\cr
6\tab 22 \tab Jönköping\cr
7\tab 23 \tab  Kronoberg\cr
8\tab 24,25 \tab  Kalmar\cr
9\tab 26 \tab  Gotland\cr
10\tab 27 \tab  Blekinge\cr
12\tab 28,30,41   \tab  Skåne\cr
13\tab 42 \tab  Halland\cr
14\tab 50,51,52,53   \tab  Västra Götaland\cr
17\tab 54 \tab  Värmland\cr
18\tab 55 \tab  Örebro\cr
19\tab 56 \tab  Västmanland\cr
20\tab 57 \tab  Dalarna\cr
21\tab 61 \tab  Gävleborg\cr
22\tab 62 \tab  Västernorrland\cr
23\tab 63 \tab  Jämtland\cr
24\tab 64 \tab  Västerbotten\cr
25\tab 65 \tab  Norrbotten\cr
-\tab 91,92,93,94,95,96 \tab  Övriga/privata - region\cr
NA\tab NA \tab  Uppgift saknas
}



If language = c("sv", "en") the following applies to arguments textBeforeSubtitle, textAfterSubtitle, comment, regionLabel, label in list varOther: if there are two values the first is used in the Swedish version and the second in the English version. If there is only one value this is recycled in both versions.
The following applies to argument outcomeTitle, description: the arguments should be given in a list, the first listargument is used in the Swedish version and the second in the English version. The Swedish title(s) will be recycled if English is missing.
The following applies to arguments outcome, geoUnitsHospital, geoUnitsCounty, geoUnitsRegion, period, var in list varOther: in the English version the variable name with the suffix _en (for example "outcome_en") will be used if this exists and otherwise the Swedish variable name will be recycled.
}
\examples{
ind1 <- rccShiny2(
  data = rccShinyData,
  folder = "Indikator1",
  outcome = paste0("outcome",1:3),
  outcomeTitle = c("Dikotom", "Kontinuerlig", "Kategorisk"),
  description = c("Har beskrivs indikatorn.","Viktig information!","Information om variabler etc."),
  varOther = list(
    list(
      var = "age",
      label = "Alder vid diagnos"
    ),
    list(
      var = "stage",
      label = "Stadium",
      choices = c("I", "II"),
      selected = "I",
      multiple = TRUE,
      showInTitle = TRUE
    )
  ),
  funnelplot = TRUE
)
\dontrun{
library(shiny)
runApp("./sv/Indikator1")

cat(ind1) # displays the html link that can be used in index.html
}

# For Swedish/English version
rccShinyData$outcome1_en <- rccShinyData$outcome1
rccShiny2(
  language = c("sv", "en"),
  data = rccShinyData,
  folder = "Indikator2",
  outcome = "outcome1",
  outcomeTitle = list("Kontaktsjukskoterska", "Contact nurse"),
  textBeforeSubtitle = c("Nagot pa svenska","Something in English"),
  description = list(
    c("Har beskrivs indikatorn.","Viktig information!","Information om variabler etc."),
    c("Description of the indicator","Important information!","Information on variables etc.")
  ),
  varOther = list(
    list(
      var = "age",
     label = c("Alder vid diagnos","Age at diagnosis"),
      choices = c(0,120)
    )
  ),
  targetValues = c(95,99)
)
}
\author{
Fredrik Sandin, RCC Uppsala-Örebro
}
