% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccShinyApp.R
\name{rccShinyCheckData}
\alias{rccShinyCheckData}
\title{Checks input to rccShiny}
\usage{
rccShinyCheckData(optionsList = NULL)
}
\description{
internal function.
}
\author{
Fredrik Sandin, RCC Uppsala-Örebro
}
