% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rccShinyCounties}
\alias{rccShinyCounties}
\title{County dataset}
\usage{
rccShinyCounties(language = "sv", lkf = FALSE)
}
\description{
internal function that creates dataset with numeric codes (given by user) and corresponding names (printed in output) for county/county of residence.
}
\author{
Fredrik Sandin, RCC Uppsala-Örebro
}
