% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rccShinyDecimals}
\alias{rccShinyDecimals}
\title{Number of decimals}
\usage{
rccShinyDecimals()
}
\description{
internal function used by rccShinyIndTable to give number of decimals.
}
\author{
Fredrik Sandin, RCC Uppsala-Örebro
}
\keyword{internal}
