% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rccShinyGroupVariable}
\alias{rccShinyGroupVariable}
\title{Group together geoUnits}
\usage{
rccShinyGroupVariable(
  label = "sjukhus",
  otherVariables = NULL,
  otherLabels = NULL,
  optionalHospitalLabel = NULL,
  optionalCountyLabel = NULL,
  optionalRegionLabel = NULL
)
}
\description{
internal function used by server.R to simplify code.
}
\author{
Fredrik Sandin, RCC Mellansverige
}
\keyword{internal}
