% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rccShinyLevelNames}
\alias{rccShinyLevelNames}
\title{Names for choices in Level of comparison}
\usage{
rccShinyLevelNames(level = "region", language = "sv")
}
\description{
internal function used to supply text printed in Level of comparison widget.
}
\author{
Fredrik Sandin, RCC Mellansverige
}
\keyword{internal}
