% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rccShinyNoObservationsText}
\alias{rccShinyNoObservationsText}
\title{Text for no observations}
\usage{
rccShinyNoObservationsText(language = "sv")
}
\description{
internal function with text to be printed if there is no data available.
}
\author{
Fredrik Sandin, RCC Mellansverige
}
\keyword{internal}
