% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rccShinyRegionNames}
\alias{rccShinyRegionNames}
\title{Region names}
\usage{
rccShinyRegionNames(all = FALSE, sort = FALSE, language = "sv")
}
\description{
internal function that gives names of regions used by rccShinyRegionNumToText.
}
\author{
Fredrik Sandin, RCC Mellansverige
}
\keyword{internal}
