% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rccShinyRegionNumToText}
\alias{rccShinyRegionNumToText}
\title{Region variable}
\usage{
rccShinyRegionNumToText(var = NULL, language = "sv")
}
\description{
internal function that creates variable from numeric codes (given by user) to corresponding names (printed in output) for region.
}
\author{
Fredrik Sandin, RCC Uppsala-Örebro
}
