% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rccShinyTabsNames}
\alias{rccShinyTabsNames}
\title{Names for tabs}
\usage{
rccShinyTabsNames(language = "sv")
}
\description{
internal function used to supply names of the tabs.
}
\author{
Fredrik Sandin, RCC Uppsala-Örebro
}
