% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rccShinyTXT}
\alias{rccShinyTXT}
\title{Miscellaneous text}
\usage{
rccShinyTXT(language = "sv")
}
\description{
internal function used to supply text for various output.
}
\author{
Fredrik Sandin, RCC Uppsala-Örebro
}
\keyword{internal}
