% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccShiny-package.R
\docType{package}
\name{rccShiny-package}
\alias{rccShiny-package}
\alias{_PACKAGE}
\title{rccShiny: Shiny Apps for RCC}
\description{
Creates shiny apps used as a complement to the annual reports from the cancer quality registries in Sweden.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://bitbucket.org/cancercentrum/rccshiny}
}

}
\author{
\strong{Maintainer}: Fredrik Sandin \email{fredrik.sandin@rccmellan.se}

Other contributors:
\itemize{
  \item Lina Benson (\href{https://orcid.org/0000-0002-5393-0055}{ORCID}) [contributor]
  \item Erik Lindberg [contributor]
  \item John Lövrot (\href{https://orcid.org/0000-0002-9339-8059}{ORCID}) [contributor]
  \item Chenyang Zhang [contributor]
  \item RCC [copyright holder, funder]
}

}
\keyword{internal}
